more work
tried to handle left-recursive grammars, but more work will be required

In particular, I almost implemented Paull's algorithm.
A->b1|...|bn is easy; invoke the normal rule for A, but have recursive A's fail.
A->a1 A'|...|am A' is not easy; it requires changing orders.
Invoking the original rule for A, and having recursive A return a known value almost works;
but distinguishing between a returned "A a" and a returned "b" is hard, especially with result rewriting.
("bi bj" probably isn't in the original rule, but could be synthesized by this approximation...)

Also note that the corner turn algorithm is more efficient.